+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Remove `concatenation-unfinished` from list of supported extensions
+
 ## 0.10.0 (2017-03-27)
 
 * Fix invalid `Content-Disposition` header in GET requests to due mutation of `Tus::Server.opts[:disposition]`